TwitterCrawler/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorlog==6.7.0
dnspython==2.3.0
docker==6.1.3
docker-pycreds==0.4.0
exceptiongroup==1.1.1
fastapi==0.96.0
flower==1.2.0
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
humanize==4.6.0
idna==3.4
Jinja2==3.1.2
kombu==5.3.0
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.3
packaging==23.1
pandas==2.0.2
playwright==1.34.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
pydantic==1.10.8
pyee==9.0.4
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
selectolax==0.3.14
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tornado==6.3.2
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.5.3
yarl==1.9.2
